+++ mitra_adapters/src/profiles.rs
@@ -1,0 +1,16 @@
+use apx_sdk::addresses::WebfingerAddress;
+use mitra_models::profiles::types::{DbActorProfile, WebfingerHostname};
+
+pub fn profile_address(
+    instance_hostname: &str,
+    profile: &DbActorProfile,
+) -> Option<WebfingerAddress> {
+    let maybe_hostname = profile.hostname();
+    let hostname = match maybe_hostname {
+        WebfingerHostname::Remote(ref hostname) => hostname,
+        WebfingerHostname::Local => instance_hostname,
+        WebfingerHostname::Unknown => return None,
+    };
+    let address = WebfingerAddress::new_unchecked(&profile.username, hostname);
+    Some(address)
+}