+++ mitra_api/src/mastodon_api/microsyntax/mentions.rs
@@ -7,6 +7,7 @@ use mitra_activitypub::{
     authority::Authority,
     identifiers::profile_actor_url,
 };
+use mitra_adapters::profiles::profile_address;
 use mitra_models::{
     database::{DatabaseClient, DatabaseError},
     profiles::queries::get_profiles_by_accts,
@@ -25,7 +26,10 @@ const MENTION_SEARCH_SECONDARY_RE: &str = r"(?x)
     (?P<after>[\.,:;?!\)']*)$
     ";
 
-fn caps_to_acct(instance_hostname: &str, caps: &Captures) -> Option<String> {
+fn caps_to_address(
+    instance_hostname: &str,
+    caps: &Captures,
+) -> Option<WebfingerAddress> {
     let username = &caps["username"];
     let hostname = if let Ok(maybe_hostname) = caps.name("hostname")
         .map(|match_| encode_hostname(match_.as_str()))
@@ -40,15 +44,14 @@ fn caps_to_acct(instance_hostname: &str, caps: &Captures) -> Option<String> {
         username,
         &hostname,
     );
-    let acct = webfinger_address.acct(instance_hostname);
-    Some(acct)
+    Some(webfinger_address)
 }
 
 /// Finds everything that looks like a mention
 fn find_mentions(
     instance_hostname: &str,
     text: &str,
-) -> Vec<String> {
+) -> Vec<WebfingerAddress> {
     let mention_re = Regex::new(MENTION_SEARCH_RE)
         .expect("regexp should be valid");
     let mention_secondary_re = Regex::new(MENTION_SEARCH_SECONDARY_RE)
@@ -61,12 +64,12 @@ fn find_mentions(
             continue;
         };
         if let Some(secondary_caps) = mention_secondary_re.captures(&caps["mention"]) {
-            let Some(acct) = caps_to_acct(instance_hostname, &secondary_caps) else {
+            let Some(address) = caps_to_address(instance_hostname, &secondary_caps) else {
                 // Invalid mention
                 continue;
             };
-            if !mentions.contains(&acct) {
-                mentions.push(acct);
+            if !mentions.contains(&address) {
+                mentions.push(address);
             };
         };
     };
@@ -77,22 +80,27 @@ pub async fn find_mentioned_profiles(
     db_client: &impl DatabaseClient,
     instance_hostname: &str,
     text: &str,
-) -> Result<IndexMap<String, DbActorProfile>, DatabaseError> {
+) -> Result<IndexMap<WebfingerAddress, DbActorProfile>, DatabaseError> {
     let mentions = find_mentions(instance_hostname, text);
+    let mut accts = vec![];
+    for address in mentions {
+        accts.push(address.acct(instance_hostname));
+    };
     // If acct doesn't exist in database, mention is ignored
-    let profiles = get_profiles_by_accts(db_client, mentions).await?;
-    let mut mention_map: IndexMap<String, DbActorProfile> = IndexMap::new();
+    let profiles = get_profiles_by_accts(db_client, accts).await?;
+    let mut mention_map = IndexMap::new();
     for profile in profiles {
-        let acct = profile.acct.as_ref()
-            .expect("acct should be present")
-            .clone();
-        mention_map.insert(acct, profile);
+        let Some(address) = profile_address(instance_hostname, &profile) else {
+            // get_profiles_by_accts should not return profiles without address
+            return Err(DatabaseError::type_error());
+        };
+        mention_map.insert(address, profile);
     };
     Ok(mention_map)
 }
 
 pub fn replace_mentions(
-    mention_map: &IndexMap<String, DbActorProfile>,
+    mention_map: &IndexMap<WebfingerAddress, DbActorProfile>,
     instance_hostname: &str,
     authority: &Authority,
     text: &str,
@@ -108,13 +116,13 @@ pub fn replace_mentions(
             return caps[0].to_string();
         };
         if let Some(secondary_caps) = mention_secondary_re.captures(&caps["mention"]) {
-            let acct = if let Some(acct) = caps_to_acct(instance_hostname, &secondary_caps) {
-                acct
+            let address = if let Some(address) = caps_to_address(instance_hostname, &secondary_caps) {
+                address
             } else {
                 // Invalid mention
                 return caps[0].to_string();
             };
-            if let Some(profile) = mention_map.get(&acct) {
+            if let Some(profile) = mention_map.get(&address) {
                 // Replace with a link to profile.
                 // Actor URL may differ from actor ID.
                 let url = profile_actor_url(authority, profile);
@@ -153,6 +161,13 @@ mod tests {
         "@user2@server2.com copy ",
         "some text",
     );
+
+    fn find_mentions(instance_hostname: &str, text: &str) -> Vec<String> {
+        super::find_mentions(instance_hostname, text)
+            .into_iter()
+            .map(|address| address.acct(instance_hostname))
+            .collect()
+    }
 
     #[test]
     fn test_find_mentions() {
@@ -219,10 +234,10 @@ mod tests {
             },
         );
         let mention_map = IndexMap::from([
-            ("user1".to_string(), profile_1),
-            ("user_x".to_string(), profile_2),
-            ("user2@server2.com".to_string(), profile_3),
-            ("user3@xn--jxalpdlp.example".to_string(), profile_4),
+            (profile_address(INSTANCE_HOSTNAME, &profile_1).unwrap(), profile_1),
+            (profile_address(INSTANCE_HOSTNAME, &profile_2).unwrap(), profile_2),
+            (profile_address(INSTANCE_HOSTNAME, &profile_3).unwrap(), profile_3),
+            (profile_address(INSTANCE_HOSTNAME, &profile_4).unwrap(), profile_4),
         ]);
         let authority = Authority::server_unchecked(INSTANCE_URI);
         let result = replace_mentions(