+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Warn about unused parameters in configuration file.
+- Added `--user-agent` parameter to `fetch-object` command.
 
 ### Changed