+++ apx_sdk/CHANGELOG.md
@@ -8,6 +8,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Added documentation to `apx_sdk::addresses` module.
+- Implemented `Hash` for `WebfingerAddress`.
+
+## [0.22.0] - 2026-03-27
+
+### Added
+
 - Added `ed25519_public_key_to_pkcs8_pem` function.
 
 ### Changed